Update ligershark.templates.targets
Added bin and obj folders to the list of folders to skip during the robocopy.

Every time I open the solution, the bin and obj folders are created even though my project template is not set to build. This causes these folders and their contents to be copied to the VSIX. I have to remember to manually delete these folders before I build my VSIX.

Adding these two folders to the excludes also has the nice side effect that you no longer need to set you project template to no build in the solution configuration.